Split types into multiple files
