+++ types/asobo-vcockpits-instruments-airliners/html_ui/Pages/VCockpit/Instruments/Airliners/Shared/BaseAirliners.d.ts
@@ -1,0 +1,14 @@
+/// <reference path="../../../../../../../fs-base-ui/html_ui/JS/common.d.ts" />
+
+declare global {
+    namespace Airliners {
+        class BaseEICAS {
+        }
+
+        class EICASTemplateElement extends TemplateElement {
+            init(): void
+        }
+    }
+}
+
+export {};
