+++ types/fs-base-ui/html_ui/JS/Inputs.d.ts
@@ -1,0 +1,18 @@
+declare global {
+    interface Inputs {
+        getInputStatus(context: any, action: any): any;
+    }
+
+    function GetInputs(): Inputs;
+
+    enum EInputStatus {
+        idle,
+        pressed,
+        down,
+        released
+    }
+
+    function GetInputStatus(context: any, action: any): any;
+}
+
+export {};
