+++ types/fs-base-ui/html_ui/JS/NetBingMap.d.ts
@@ -1,0 +1,73 @@
+declare global {
+    class BingMapsConfig {
+        load(source: {
+            netBingHeightColor1?: any;
+            netBingAltitudeColors1?: any;
+            netBingTextureResolution?: any;
+            netBingSkyColor1?: any;
+            netBingHeightColor2?: any;
+            netBingAltitudeColors2?: any;
+            netBingSkyColor2?: any;
+            netBingHeightColor3?: any;
+            netBingAltitudeColors3?: any;
+            netBingSkyColor3?: any;
+        }, id: number): boolean;
+        heightColors: number[];
+        aspectRatio: number;
+        resolution: number;
+        clearColor: number;
+    }
+
+    enum EBingMode {
+        CURSOR = "Cursor",
+        PLANE = "Plane",
+        VFR = "Vfr",
+        HORIZON = "Horizon"
+    }
+
+    enum EBingReference {
+        SEA = "Sea",
+        PLANE = "Plane"
+    }
+
+    enum EWeatherRadar {
+        OFF = "Off",
+        TOPVIEW = "Topview",
+        HORIZONTAL = "Horizontal",
+        VERTICAL = "Vertical"
+    }
+
+    class BingMapsBinder {
+    }
+
+    class BingMapElement extends HTMLElement {
+        onListenerRegistered: () => void;
+        onListenerBinded: (binder: { friendlyName: any; is3D: boolean }, uid: any) => void;
+        updateMapImage: (uid: any, img: string) => void;
+        OnDestroy: () => void;
+        connectedCallback(): void;
+        disconnectedCallback(): void;
+        isReady(): boolean;
+        setBingId(id: any): void;
+        setMode(mode: EBingMode): void;
+        setReference(ref: EBingReference): void;
+        addConfig(config: BingMapsConfig): void;
+        setConfig(configId: number): void;
+        setParams(value: any): void;
+        showIsolines(show: any): void;
+        getIsolines(): any;
+        showWeather(mode: any, cone: any): void;
+        getWeather(): any;
+        setVisible(show: any): void;
+        is3D(): boolean;
+        updateBinding(): void;
+        updateConfig(): void;
+        updateReference(): void;
+        updatePosAndSize(): void;
+        updateVisibility(): void;
+        updateIsolines(): void;
+        updateWeather(): void;
+    }
+}
+
+export {};
