+++ types/fs-base-ui/html_ui/JS/animation/animation.d.ts
@@ -1,0 +1,7 @@
+declare global {
+    class UITimeline {
+        // TODO fill this.
+    }
+}
+
+export {};
