+++ types/fs-base-ui/html_ui/JS/debug.d.ts
@@ -1,0 +1,9 @@
+declare global {
+    // The following variables are already declared in fs-base-ui/html_ui/JS/common.js.
+    // bLiveReload = false;
+    // bAutoReloadCSS = false;
+    // bDebugListeners = false;
+    // bDebugKeyNavigation = false;
+}
+
+export {};
